feat(ui): account name from Location Name CV + in-app Help page; fix(customfields): float position; docs: de-couple from [W]-OS
ConnectLocation resolves a friendly account name from the Location Name custom value (falls back to locationId), so titles/library show the name not the raw id.

In-app Help page (/help + nav item) renders docs/MANUAL.md via a dependency-free markdown renderer (lib/utils/markdown.ts) + vite server.fs.allow for ../docs.

fix: decode GHL customField.position as float64 (GHL returns e.g. 281.25, was a 502).

docs: add docs/MANUAL.md (current+future feature reference); scrub MANUAL.md + architecture doc of all [W]-OS / vigil / internal-ecosystem references. Standalone product; MCP tools + Claude shell stay, decoupled.

+++ backend/internal/handlers/customfields.go
@@ -12,11 +12,11 @@ import (
 )
 
 type customField struct {
-	ID       string `json:"id"`
-	Name     string `json:"name"`
-	DataType string `json:"dataType,omitempty"`
-	FieldKey string `json:"fieldKey,omitempty"`
-	Position int    `json:"position,omitempty"`
+	ID       string  `json:"id"`
+	Name     string  `json:"name"`
+	DataType string  `json:"dataType,omitempty"`
+	FieldKey string  `json:"fieldKey,omitempty"`
+	Position float64 `json:"position,omitempty"` // GHL returns fractional positions (e.g. 281.25), not ints
 }
 
 // ListCustomFields returns a location's custom-field schema
